preproc

+++ preprocess_data.py
@@ -9,6 +9,6 @@ n_actions=len(set(data["action"]))
 episode_set=set()
 data=data.select([z for z in range(20)])
 data=data.map(lambda x :{"image": image_processor.preprocess( x["image"])[0]})
-data=data.map(lambda x: {"action":F.one_hot(torch.Tensor(x["action"]),n_actions)})
+data=data.map(lambda x: {"action":F.one_hot(torch.Tensor(x["action"]).long(),n_actions)})
 
 data.push_to_hub("jlbaker361/sonic-vae-preprocessed")
